move DictionaryQueryEntry to @subql/types
